+++ CHANGELOG.md
@@ -28,6 +28,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AbstractNode class
 - NodeWithDefaultLabel class
 - field 'title' with tooltip text in RiskNode primitive properties
+- new RiskAppetiteGoal and RiskAppetiteRisk primitives instead of RiskAppetite
+
+### Changed
+
+- version of sdk
+- node titles
 
 ## [0.2.0]